extract container functino to trait
