+++ src/Attribute/AbstractAttribute.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace Reinfi\DependencyInjection\Attribute;
 
-use Psr\Container\ContainerInterface;
-use Laminas\ServiceManager\AbstractPluginManager;
+use Reinfi\DependencyInjection\Injection\ContainerTrait;
 use Reinfi\DependencyInjection\Injection\InjectionInterface;
 
 /**
@@ -13,13 +12,5 @@ use Reinfi\DependencyInjection\Injection\InjectionInterface;
  */
 abstract class AbstractAttribute implements InjectionInterface
 {
-    protected function determineContainer(ContainerInterface $container): ContainerInterface
-    {
-        if ($container instanceof AbstractPluginManager) {
-            return $container
-                ->getServiceLocator();
-        }
-
-        return $container;
-    }
+    use ContainerTrait;
 }
